requirements.txt
langchain
langchain-openai
python-dotenv
openai
streamlit
icecream